Add Side to match, add radiant & dire methods to match
Add get_JSON_any method to client
Move player and side definitions to own files
Change folder structure for better organization
Update README accordingly
Update .travis.yml to only check master branch

Basic definitions & sub-definitions pertaining to
classes are now in their own folders. Much cleaner
this way, and looks better in dota.cr requirements
calls.

+++ src/dota/api/basic/basic_player.cr
@@ -3,7 +3,7 @@ module Dota
     class BasicPlayer
       JSON.mapping(
         account_id: {type: Int64, nilable: true},
-        player_slot: Int8,
+        player_slot: Int16,
         hero_id: Int8
       )
     end
